Sometimes we miss removable CheckInBounds https://bugs.webkit.org/show_bug.cgi?id=200018
Reviewed by Saam Barati.

JSTests:

* microbenchmarks/typed-array-sum.js: Added.
(doTest):

Source/JavaScriptCore:

We failed to remove the CheckInBounds bounds because we did not see that the index was nonnegative. This is because we do not see the relationship between the two
separate zero constants that appear in the IR for the given test case. This patch re-adds the hack to de-duplicate m_zero that was removed in
<https://trac.webkit.org/changeset/241228/webkit>.

* dfg/DFGIntegerRangeOptimizationPhase.cpp:


Identifier: 213883@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@247724 268f45cc-cd09-0410-ab3c-d52691b4dbfc