Multiple context menu actions broken for YouTube videos https://bugs.webkit.org/show_bug.cgi?id=199999
Reviewed by Eric Carlson.

Source/WebCore:

Do not include CopyMediaLinkItem and OpenMediaInNewWindowItem to the context menu when media URL is not
downloadable or the request can't be handled.

* page/ContextMenuController.cpp:
(WebCore::ContextMenuController::populate):

Tools:

Add a test case to check copy link address, open in new window and download options are not included in the
context menu for non-downloadable media.

* TestWebKitAPI/Tests/WebKitGtk/TestContextMenu.cpp:
(writeNextChunk):
(serverCallback):
(testContextMenuLiveStream):
(beforeAll):
(afterAll):

Identifier: 214017@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@247901 268f45cc-cd09-0410-ab3c-d52691b4dbfc